+++ src/main/java/com/gajava/library/service/RecordServiceImpl.java
@@ -1,6 +1,7 @@
 package com.gajava.library.service;
 
 import com.gajava.library.exception.EntityNotFoundException;
+import com.gajava.library.exception.InvalidArgumentsException;
 import com.gajava.library.exception.SaveEntityException;
 import com.gajava.library.model.Book;
 import com.gajava.library.model.Reader;
@@ -13,6 +14,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -23,6 +25,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Record save(final Record record) {
+        if (Objects.isNull(record)) {
+            throw new InvalidArgumentsException("The record cannot be null");
+        }
         log.info("Try to save record");
         final Optional<Record> recordCreated = Optional.of(recordRepository.save(record));
         return recordCreated.orElseThrow(() -> new SaveEntityException("Record"));
@@ -30,6 +35,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Record findById(final Long id) {
+        if (Objects.isNull(id)) {
+            throw new InvalidArgumentsException("The record id cannot be null");
+        }
         log.info("Trying to find a record by id " + id);
         return recordRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Record", id));
     }
@@ -37,23 +45,32 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public void delete(final Long id) {
+        if (Objects.isNull(id)) {
+            throw new InvalidArgumentsException("The record id cannot be null");
+        }
         log.info("Trying to remove record with id " + id);
         try {
             recordRepository.deleteById(id);
             log.info("Deleting the record from id " + id + " successfully");
         } catch (EmptyResultDataAccessException e) {
-            throw new EntityNotFoundException("Reader", id);
+            throw new EntityNotFoundException("Record", id);
         }
     }
 
     @Override
     public Page<Record> findAll(final Pageable pageable) {
+        if (Objects.isNull(pageable)) {
+            throw new InvalidArgumentsException("The pageable cannot be null");
+        }
         log.info("Looking for all the records");
         return recordRepository.findAll(pageable);
     }
 
     @Override
     public Page<Record> findAllByReader(final Reader reader, final Pageable pageable) {
+        if (Objects.isNull(reader) || Objects.isNull(pageable)) {
+            throw new InvalidArgumentsException("The reader or pageable cannot be null");
+        }
         log.info("Looking for a record by reader id = " + reader.getId());
         final Optional<Page<Record>> records = Optional.ofNullable(
                 recordRepository.findAllByReader(reader, pageable));
@@ -62,6 +79,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Page<Record> findAllByBook(final Book book, final Pageable pageable) {
+        if (Objects.isNull(book) || Objects.isNull(pageable)) {
+            throw new InvalidArgumentsException("The book or pageable cannot be null");
+        }
         log.info("Looking for a record by book id = " + book.getId());
         final Optional<Page<Record>> records = Optional.ofNullable(
                 recordRepository.findAllByBook(book, pageable));
@@ -70,6 +90,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Page<Record> findAllByDateValidReturnIsNull(final Pageable pageable) {
+        if (Objects.isNull(pageable)) {
+            throw new InvalidArgumentsException("The pageable cannot be null");
+        }
         log.info("Looking for a record by book on hands");
         final Optional<Page<Record>> records = Optional.ofNullable(
                 recordRepository.findAllByDateValidReturnIsNull(pageable));
@@ -78,6 +101,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Page<Record> findAllByDateValidReturnIsNotNull(final Pageable pageable) {
+        if (Objects.isNull(pageable)) {
+            throw new InvalidArgumentsException("The pageable cannot be null");
+        }
         log.info("Looking for a record by book returned");
         final Optional<Page<Record>> records = Optional.ofNullable(
                 recordRepository.findAllByDateValidReturnIsNotNull(pageable));
@@ -86,6 +112,9 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public Record findRecordByReaderIdAndBookId(final Long readerId, final Long bookId) {
+        if (Objects.isNull(readerId) || Objects.isNull(bookId)) {
+            throw new InvalidArgumentsException("The book id or reader id cannot be null");
+        }
         log.info("Looking for a record by reader id " + readerId + " and book id " + bookId);
         final Optional<Record> record = Optional.ofNullable(
                 recordRepository.findFirstByReaderIdAndBookIdAndDateValidReturnIsNull(readerId, bookId));
